Adding Azure Core Files

+++ wp-content/themes/listable/job_manager/content-job_listing.php
@@ -40,18 +40,20 @@ if ( true === $listing_is_featured ) {
 	$listing_classes .= '  is--featured';
 }
 
+$card_image = listable_get_post_image_src( $post->ID, 'listable-card-image' );
+
 $listing_classes = apply_filters( 'listable_listing_archive_classes', $listing_classes, $post ); ?>
 <div class="grid__item">
 	<article class="<?php echo esc_attr( $listing_classes ); ?>" itemscope itemtype="http://schema.org/LocalBusiness"
 	         data-latitude="<?php echo esc_attr( get_post_meta( $post->ID, 'geolocation_lat', true ) ); ?>"
 	         data-longitude="<?php echo esc_attr( get_post_meta( $post->ID, 'geolocation_long', true ) ); ?>"
-	         data-img="<?php echo esc_attr( listable_get_post_image_src( $post->ID, 'listable-card-image' ) ); ?>"
+	         data-img="<?php echo esc_attr( $card_image ); ?>"
 	         data-permalink="<?php echo esc_attr( get_the_job_permalink() ); ?>"
 	         data-categories="<?php echo esc_attr( $termString ); ?>"
 		<?php echo $data_output; ?> >
 		<a href="<?php the_job_permalink(); ?>">
 			<aside class="card__image"
-			       style="background-image: url(<?php echo listable_get_post_image_src( $post->ID, 'listable-card-image' ); ?>);">
+			       style="background-image: url(<?php echo esc_url( $card_image ? str_replace( 'https://', '//', $card_image ) : apply_filters( 'listing_card_placeholer', '') ); ?>);">
 				<?php if ( true === $listing_is_featured ): ?>
 					<span class="card__featured-tag"><?php esc_html_e( 'Featured', 'listable' ); ?></span>
 				<?php endif; ?>
@@ -64,7 +66,7 @@ $listing_classes = apply_filters( 'listable_listing_archive_classes', $listing_c
 			<div class="card__content">
 				<h2 class="card__title" itemprop="name"><?php
 					echo get_the_title();
-					if ( $listing_is_claimed === '1' ) {
+					if ( $listing_is_claimed === '1' && function_exists('wpjmcl_init')) {
 						echo '<span class="listing-claimed-icon">';
 						get_template_part( 'assets/svg/checked-icon-small' );
 						echo '<span>';
